requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.1
django-cors-headers==4.3.1
django-otp==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.2
sqlparse==0.5.0
timedelta==2020.12.3
typing_extensions==4.11.0
urllib3==2.2.1
